tests/wpt/tests/webauthn/createcredential-prf.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>navigator.credentials.create() prf extension tests with authenticator support</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src=helpers.js></script>
<body></body>
<script>
standardSetup(function() {
  "use strict";

  promise_test(async t => {
    const credential = await createCredential({
      options: {
        publicKey: {
          extensions: {
            prf: {},
          },
        },
      },
    });
    assert_true(credential.getClientExtensionResults().prf.enabled);
  }, "navigator.credentials.create() with prf requested");

  promise_test(async t => {
    const promise = createCredential({
      options: {
        publicKey: {
          extensions: {
            prf: {evalByCredential: {"Zm9v": {first: new Uint8Array([1,2,3,4]).buffer}}},
          },
        },
      },
    });
    return promise_rejects_dom(t, "NotSupportedError", promise);
  }, "navigator.credentials.create() with nonsensical evalByCredential");
}, {
  protocol: "ctap2_1",
  extensions: ["prf"],
});
</script>
